sec_logistics/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.28
supabase==2.3.5
pydantic==2.6.3
python-dotenv==1.0.1
aiosqlite==0.20.0
